refactor: home battle controller 추가 및 bind 파일 위치 변경
