+++ lib/main.dart
@@ -5,19 +5,11 @@ import 'package:intl/date_symbol_data_local.dart';
 import 'package:get/get.dart';
 import 'package:kakao_flutter_sdk/kakao_flutter_sdk.dart';
 import 'package:poorlex/bind/friends.dart';
-import 'package:poorlex/bind/modify_battle.dart';
-import 'package:poorlex/controller/api.dart';
-import 'package:poorlex/controller/battle.dart';
+import 'package:poorlex/bind/battle/modify_battle.dart';
+import 'package:poorlex/bind/root_bind.dart';
 import 'package:poorlex/controller/hive_box.dart';
-import 'package:poorlex/controller/image_picker.dart';
-import 'package:poorlex/controller/layout.dart';
-import 'package:poorlex/controller/user.dart';
 import 'package:poorlex/libs/theme.dart';
 import 'package:poorlex/middleware/auth_middleware.dart';
-import 'package:poorlex/provider/battles_provider.dart';
-import 'package:poorlex/provider/expenditures_povider.dart';
-import 'package:poorlex/provider/login_provider.dart';
-import 'package:poorlex/provider/member_provider.dart';
 import 'package:poorlex/screen/battle/modify_battle_detail.dart';
 import 'package:poorlex/screen/friends/friends.dart';
 import 'package:poorlex/screen/my/my_notification.dart';
@@ -28,46 +20,6 @@ import 'package:poorlex/screen/my/my_expense_input.dart';
 import 'package:poorlex/screen/my/my_option.dart';
 import 'package:poorlex/screen/my/my_profile.dart';
 import 'package:poorlex/screen/battle/battle_create.dart';
-
-class Bind extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(() => LoginProvider());
-    Get.lazyPut(() => MemberProvider());
-    Get.lazyPut(() => ExpendituresProvider());
-
-    Get.put(LayoutController(), permanent: true);
-    Get.put(ApiController(), permanent: true);
-    Get.put(
-      UserController(
-        memberProvider: Get.find(),
-        expendituresProvider: Get.find(),
-        loginProvider: Get.find(),
-      ),
-      permanent: true,
-    );
-    Get.put(ImagePickerController(), permanent: true);
-  }
-}
-
-class BattleBind extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(
-      () => BattlesProvider(
-        user: Get.find(),
-      ),
-    );
-    Get.put(
-      BattleController(
-        battlesProvider: Get.find(),
-        layout: Get.find(),
-        imagePickerController: Get.find(),
-      ),
-      permanent: true,
-    );
-  }
-}
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -88,7 +40,7 @@ void main() async {
 
   runApp(
     GetMaterialApp(
-      initialBinding: Bind(),
+      initialBinding: RootBind(),
       initialRoute: '/',
       theme: ThemeData(
         fontFamily: 'NeoDunggeunmoPro-Regular',
@@ -134,7 +86,7 @@ void main() async {
                 'https://horse-whitefish-a82.notion.site/3623856463694fd2b9f38806b8cf507e?pvs=4',
           ),
         ),
-        GetPage(name: '/my/profile', page: () => MyProfile(), binding: Bind()),
+        GetPage(name: '/my/profile', page: () => MyProfile()),
         GetPage(
           name: '/my/expense-input',
           page: () => MyExpenseInputPage(),
