+++ lib/screen/battle/modify_battle_detail.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
 import 'package:poorlex/controller/modify_battle.dart';
 import 'package:poorlex/libs/theme.dart';
