+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: poorlex
 description: poorlex project
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 0.0.3+02
+version: 0.0.4+00
 
 environment:
   sdk: ">=3.0.6 <4.0.0"
